+++ pybricks/experimental/pb_module_experimental.c
@@ -9,6 +9,9 @@
 #include "py/obj.h"
 #include "py/runtime.h"
 
+#include <pybricks/util_mp/pb_obj_helper.h>
+#include <pybricks/util_mp/pb_kwarg_helper.h>
+
 #include <pybricks/experimental.h>
 #include <pybricks/robotics.h>
 
@@ -46,6 +49,46 @@ STATIC mp_obj_t mod_experimental_pthread_raise(mp_obj_t thread_id_in, mp_obj_t e
 STATIC MP_DEFINE_CONST_FUN_OBJ_2(mod_experimental_pthread_raise_obj, mod_experimental_pthread_raise);
 #endif // PYBRICKS_HUB_EV3BRICK
 
+
+// pybricks.experimental.hello_world
+STATIC mp_obj_t experimental_hello_world(size_t n_args, const mp_obj_t *pos_args, mp_map_t *kw_args) {
+    PB_PARSE_ARGS_FUNCTION(n_args, pos_args, kw_args,
+        // Add up to 8 arguments below, all separated by one comma.
+        // You can choose from:
+        //  PB_ARG_REQUIRED(name), This is a required argument.
+        //  PB_ARG_DEFAULT_INT(name, value), Keyword argument with default int value.
+        //  PB_ARG_DEFAULT_OBJ(name, value), Keyword argument with default MicroPython object.
+        //  PB_ARG_DEFAULT_QSTR(name, value), Keyword argument with default string value, without quotes.
+        //  PB_ARG_DEFAULT_FALSE(name), Keyword argument with default False value.
+        //  PB_ARG_DEFAULT_TRUE(name), Keyword argument with default True value.
+        //  PB_ARG_DEFAULT_NONE(name), Keyword argument with default None value.
+        PB_ARG_REQUIRED(foo),
+        PB_ARG_DEFAULT_NONE(bar));
+
+    // This function can be used in the following ways:
+    // from pybricks.experimental import hello_world
+    // y = hello_world(5)
+    // y = hello_world(foo=5)
+    // y = hello_world(5, 6)
+    // y = hello_world(foo=5, bar=6)
+
+    // All input arguments are available as MicroPython objects with _in post-fixed to the name.
+    // Use MicroPython functions or Pybricks helper functions to unpack them into C types:
+    mp_int_t foo = pb_obj_get_int(foo_in);
+
+    // You can check if an argument is none.
+    if (bar_in == mp_const_none) {
+        // Example of how to print.
+        mp_printf(&mp_plat_print, "Bar was not given. Foo is: %d\n", foo);
+    }
+
+    // Example of returning an object. Here we return the square of the input argument.
+    return mp_obj_new_int(foo * foo);
+}
+// See also experimental_globals_table below. This function object is added there to make it importable.
+STATIC MP_DEFINE_CONST_FUN_OBJ_KW(experimental_hello_world_obj, 0, experimental_hello_world);
+
+
 STATIC const mp_rom_map_elem_t experimental_globals_table[] = {
     #if PYBRICKS_HUB_EV3BRICK
     { MP_ROM_QSTR(MP_QSTR___name__), MP_ROM_QSTR(MP_QSTR_experimental_c) },
@@ -54,6 +97,7 @@ STATIC const mp_rom_map_elem_t experimental_globals_table[] = {
     #else
     { MP_ROM_QSTR(MP_QSTR___name__), MP_ROM_QSTR(MP_QSTR_experimental) },
     #endif // PYBRICKS_HUB_EV3BRICK
+    { MP_ROM_QSTR(MP_QSTR_hello_world), MP_ROM_PTR(&experimental_hello_world_obj) },
 };
 STATIC MP_DEFINE_CONST_DICT(pb_module_experimental_globals, experimental_globals_table);